% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{assert_language_in_namespace}
\alias{assert_language_in_namespace}
\title{Assert a language is available in a namespace}
\usage{
assert_language_in_namespace(language, ns)
}
\arguments{
\item{language}{two-character language code, e.g. \code{en}}

\item{ns}{namespace used from \code{translations_import}}
}
\value{
No return value.
}
\description{
Assert a language is available in a namespace
}
\examples{
\dontrun{
assert_language_in_namespace('en', 'r18n')
}
}
\keyword{internal}
