% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{assert_namespace}
\alias{assert_namespace}
\title{Assert a namespace has been initialized}
\usage{
assert_namespace(ns)
}
\arguments{
\item{ns}{namespace used from \code{translations_import}}
}
\value{
No return value.
}
\description{
Assert a namespace has been initialized
}
\keyword{internal}
