% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caller.R
\name{caller}
\alias{caller}
\title{Find calling package}
\usage{
caller(shift = 0L)
}
\arguments{
\item{shift}{integer if need to shift the number of parent environments}
}
\value{
package name or \code{NULL}
}
\description{
Find calling package
}
