% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_po}
\alias{generate_po}
\title{Generate dummy PO file content}
\usage{
generate_po(
  copyright_holder,
  package_name,
  package_date,
  language_team,
  language
)
}
\arguments{
\item{copyright_holder}{string}

\item{package_name}{string}

\item{package_date}{string}

\item{language_team}{string}

\item{language}{string}
}
\value{
string
}
\description{
Generate dummy PO file content
}
\keyword{internal}
