% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po.R
\name{getlines}
\alias{getlines}
\title{Extract lines from a character vector with the provided prefix}
\usage{
getlines(lines, prefix)
}
\arguments{
\item{lines}{string}

\item{prefix}{string}
}
\description{
Extract lines from a character vector with the provided prefix
}
\keyword{internal}
