% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po.R
\name{po_folder}
\alias{po_folder}
\title{Returns the i18n folder of an installed R package}
\usage{
po_folder(package)
}
\arguments{
\item{package}{string}
}
\value{
path
}
\description{
Returns the i18n folder of an installed R package
}
