% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po.R
\name{po_read}
\alias{po_read}
\title{Parse \code{PO} file}
\usage{
po_read(file)
}
\arguments{
\item{file}{PO file path}
}
\value{
list
}
\description{
This is a quick and dirty \code{PO} file parser that should be
replaced with a proper library when needed. For the record,
\pkg{poio} looks promising, but I don't like bringing in
\pkg{dplyr} dependency, so will just go with this home-brewed
solution for now.
}
\details{
Potential issues: double quotes in strings, plural forms.
}
\keyword{internal}
