% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{test_translate}
\alias{test_translate}
\title{Test if translation works for a language by looking up a term}
\usage{
test_translate(language = "en")
}
\arguments{
\item{language}{two-character language code, e.g. \code{en}}
}
\value{
string
}
\description{
Test if translation works for a language by looking up a term
}
