% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translatable}
\alias{translatable}
\alias{T}
\title{Dummy function to mark text to be added to the list of terms to be translated}
\usage{
translatable(text, comment_to_translator = NA_character_)
}
\arguments{
\item{text}{string to be translated}

\item{comment_to_translator}{optional free text to be added to the
\code{po} file for the translator}
}
\value{
\code{text}
}
\description{
Dummy function to mark text to be added to the list of terms to be translated
}
