% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate string}
\usage{
translate(text, language = translate_get_language(ns), ns = caller())
}
\arguments{
\item{text}{string to be translated}

\item{language}{target language for the translation}

\item{ns}{namespace used from \code{translations_import}}
}
\value{
string
}
\description{
Translate string
}
\note{
This is usually need to be called around \code{translatable}
    or \code{T}, as extracting all text withing \code{translate}
    would not be possible due to dynamic values to be translated
    (e.g. when text to be translated is stored in a variable).
}
