% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate_get_language}
\alias{translate_get_language}
\title{Get current translation language}
\usage{
translate_get_language(ns)
}
\arguments{
\item{ns}{namespace used from \code{translations_import}}
}
\value{
string
}
\description{
Get current translation language
}
