% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate_set_language}
\alias{translate_set_language}
\title{Set translation language}
\usage{
translate_set_language(language, ns)
}
\arguments{
\item{ns}{namespace used from \code{translations_import}}
}
\value{
No return value.
}
\description{
Set translation language
}
