% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translated_languages}
\alias{translated_languages}
\title{List translated languages based on the list of PO files in the folder}
\usage{
translated_languages(ns)
}
\arguments{
\item{ns}{namespace used from \code{translations_import}}
}
\value{
character vector of language codes
}
\description{
List translated languages based on the list of PO files in the folder
}
