% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translations}
\alias{translations}
\title{Return all translations for a language}
\usage{
translations(language, ns)
}
\arguments{
\item{ns}{namespace used from \code{translations_import}}
}
\value{
list
}
\description{
Return all translations for a language
}
