% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translations_cleanup}
\alias{translations_cleanup}
\title{Drop all translations and settings for a namespace}
\usage{
translations_cleanup(ns)
}
\arguments{
\item{ns}{namespace used from \code{translations_import}}
}
\value{
No return value.
}
\description{
Drop all translations and settings for a namespace
}
