% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{translations_generate}
\alias{translations_generate}
\title{Generate \code{PO} files for the package based on \code{translate} calls}
\usage{
translations_generate(
  folder,
  copyright_holder = NA_character_,
  language_team = NA_character_,
  languages
)
}
\arguments{
\item{folder}{path to R package}

\item{copyright_holder}{name of the organization or person owning the rights over the PO files}

\item{language_team}{name and email address of the team}

\item{languages}{optional character vector of language codes to translate (use to add new languages)}
}
\description{
Generate \code{PO} files for the package based on \code{translate} calls
}
\note{
System dependency of \code{gettext}.
}
