% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translations_import}
\alias{translations_import}
\title{Load PO files and cache in the package namespace}
\usage{
translations_import(folder, ns = caller())
}
\arguments{
\item{folder}{path to root folder of language files, where to find
the \code{po} file for \code{language}, e.g. \code{en.po}}

\item{ns}{namespace under the content of the PO files will be stored}
}
\value{
No return value.
}
\description{
Load PO files and cache in the package namespace
}
\examples{
\dontrun{
translations_import(system.file('i18n', package = 'r18r'), 'r18r')
translate_set_language('en', 'r18r')
translate_get_language('r18r')
translations('en', 'r18r')
translate('Text to be translated', ns = 'r18r')
translations_cleanup('r18r')
}
}
